requirements.txt
# Core dependencies
networkx>=2.8.0
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
seaborn>=0.11.0

# Configuration management
PyYAML>=6.0

# Optional: For enhanced CLI
typer>=0.7.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
ruff>=0.0.100

# For potential Streamlit demo
streamlit>=1.20.0
